Add remaining encoder commands and split usage tracking from validation
Most encoder commands are now implemented. set_viewport is not yet,
and the texture copy commands need row_pitch conversion.

Usage tracking has been moved into the command encoding and validation
is non-existant (right now). This is a shift away from Dawn as it
combines usage tracking and validation. The idea is that by splitting
them as separate concerns, that the code will be easier to follow.