Prevent the triangle sample from crashing when minimized
When the swapchain is rebuilt after minimizing a window
(or resizing it so that one dimension is 0) a validation
error is reported. Zero isn't a valid width or height for
when creating the swapchain, and the surface reports the
currentExtent with and height as zero. Although the error
is reported, the swapchain creation does not fail.

However, when multisampling is enabled, the output texture
creation fails and the app crashes. This adds a check only
to the multisample example.