Select min image size for framebuffer; FIFO present mode for low power
Winit versions prior to eventloop-2.0 (which is not yet released)
run the event loop in a secondary thread. When rendering, the last
window resize event received on the render thread may not match
the actual surface dimensions during window resize. This creates
problems for swapchain and framebuffer creation. Vulkan emits
a validation message when any image attachment is smaller than
the framebuffer dimensions, and will eventually result in an error.

In addition, the LowPower power preference now causes swapchain
present mode to always use FIFO (v-sync).