Use vec instead of slices in most descriptors
With the exception of RenderPassDescriptor (and friends), all
descriptors no longer have lifetime bounds. This should allow
for easier serialization and management of those descriptors.

The RenderPassDescriptor(s) need to be constructed per frame
in any case. Allowing for those to be cloned out of scope
might encourage attempting to re-use them, which may have an
unintended side effect of holding references to the swap chain
images longer than intended.