Mitigate segfault with NVIDA driver when rendering from second thread
The NVIDIA driver sometimes segfaults after swapchain recreation
from a secondary thread during window resize. This issue has
not been observed with the intel drivers for integrated UHD 630.

Waiting for the device to become idle asserts that all images
from the existing swapchain are no longer in use. Any previously
re-created swapchain is now always purged.

This issues does not seem to manifest when both the eventloop
and rendering is on a single thread. It's not clear if the fix
here fully resolves the problem or if error condition is just
much less likely now.