Change all offsets to usize
Normalizing on `usize` is more consistent and idiomatic. This
eliminates the need to cast from `usize` to `u32` or `u64`.

The `offset_of!` macro has also been removed and replaced with
the memoffset crate.

Added minor optimization for command buffer recording to reduce
allocations (use SmallVec instead of Vec for offsets and buffers).